% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_inline.R
\name{css_inline}
\alias{css_inline}
\title{Perform CSS inlining with HTML}
\usage{
css_inline(html)
}
\arguments{
\item{html}{HTML provided as a string.}
}
\value{
A character vector of length one which contains HTML text.
}
\description{
Given an HTML string, the \code{css_inline()} function will render that into HTML
with CSS inlined into the \code{style} attribute of HTML tags.
}
